requirements.txt
numpy
pandas
scikit-learn
scikit-image
Pillow
matplotlib
